Rename CKEditor component file to PascalCase
