+++ src/Ckeditor.vue
